Post return after validation
we want to post a return after customer_return object validation
to eventually add errors to customer_return and inhibite save.

To make this possible we had to change the proxy create_on_shipwire method
to support objects (like CustomerReturn) that are not persisted before
the shipwire request is made.
So we added update_shipwire_id method that checks if the object is persisted
or not.