Don't create the CRUD routes for shipments
There is no need to create all the CRUD routes for shipments, the
`solidus_api` gem adds only the `create` and the `update` routes.
Adding also the `show` leads to a conflict with the `mine` route because
in the url `api/shipments/mine`, the `mine` part is treated as the id of
the shipment for the `show` action and not as a standalone route.